Rename presentation package to compose
