+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/orders/details/OrderDetailsScreen.kt
@@ -30,11 +30,11 @@ import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.wear.compose.material.TimeText
 import androidx.wear.tooling.preview.devices.WearDevices
 import com.woocommerce.android.R
-import com.woocommerce.android.presentation.component.ErrorScreen
-import com.woocommerce.android.presentation.component.LoadingScreen
-import com.woocommerce.android.presentation.theme.WooColors
-import com.woocommerce.android.presentation.theme.WooTheme
-import com.woocommerce.android.presentation.theme.WooTypography
+import com.woocommerce.android.compose.component.ErrorScreen
+import com.woocommerce.android.compose.component.LoadingScreen
+import com.woocommerce.android.compose.theme.WooColors
+import com.woocommerce.android.compose.theme.WooTheme
+import com.woocommerce.android.compose.theme.WooTypography
 import com.woocommerce.android.ui.orders.FormatOrderData.OrderItem
 import com.woocommerce.android.ui.orders.FormatOrderData.ProductItem
