+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/stats/StoreStatsScreen.kt
@@ -31,11 +31,11 @@ import androidx.wear.compose.material.Text
 import androidx.wear.compose.material.TimeText
 import androidx.wear.tooling.preview.devices.WearDevices
 import com.woocommerce.android.R
-import com.woocommerce.android.presentation.component.ErrorScreen
-import com.woocommerce.android.presentation.component.LoadingScreen
-import com.woocommerce.android.presentation.theme.WooColors
-import com.woocommerce.android.presentation.theme.WooTheme
-import com.woocommerce.android.presentation.theme.WooTypography
+import com.woocommerce.android.compose.component.ErrorScreen
+import com.woocommerce.android.compose.component.LoadingScreen
+import com.woocommerce.android.compose.theme.WooColors
+import com.woocommerce.android.compose.theme.WooTheme
+import com.woocommerce.android.compose.theme.WooTypography
 
 @Composable
 fun StoreStatsScreen(viewModel: StoreStatsViewModel) {
